# openapi-schema.yaml
openapi: 3.0.0
info:
  title: Tarkov Casino API
  version: 1.0.0
  description: |
    API for Tarkov Casino gaming platform
    
    Note: This schema is used by Cloudflare API Shield for request validation.
    The production server is protected by Cloudflare with schema validation,
    rate limiting, and security headers.
servers:
  - url: https://tarkov.juanis.cool/api
    description: Production server (protected by Cloudflare API Shield)
...
components:
  securitySchemes:
    AppwriteSession:
      type: apiKey
      in: header
      name: X-Appwrite-Session
      description: Appwrite session secret
    AppwriteUser:
      type: apiKey
      in: header
      name: X-Appwrite-User-Id
      description: Appwrite user ID
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
paths:
...
  /games/roulette/bet:
    post:
      tags:
        - games
      summary: Place roulette bet
      security:
        - AppwriteSession: []
        - AppwriteUser: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - betType
                - betValue
              properties:
                amount:
                  type: number
                  minimum: 1
                  maximum: 10000
                  description: Bet amount
                betType:
                  type: string
                  minLength: 1
                  maxLength: 20
                  description: Bet type
                betValue:
                  oneOf:
                    - type: number
                    - type: string
                  description: Bet value
      responses:
        '200':
          description: Bet placed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  game_result:
                    type: object
                  bet_amount:
                    type: number
                  win_amount:
                    type: number
                  net_result:
                    type: number
                  new_balance:
                    type: number
                  game_id:
                    type: string
        '400':
          description: Invalid bet or insufficient balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Case opening game
...
  /games/cases/{caseTypeId}:
    get:
      tags:
        - games
      summary: Get case type details
      parameters:
        - name: caseTypeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Case type details
        '404':
          description: Case type not found
  /games/cases/open:
    post:
      tags:
        - games
      summary: Open a case
      security:
        - AppwriteSession: []
        - AppwriteUser: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - caseTypeId
              properties:
                caseTypeId:
                  type: string
                  minLength: 1
                  description: Case type ID
                previewOnly:
                  type: boolean
                  default: false
                  description: Preview only without opening
                requestId:
                  type: string
                  description: Request deduplication ID
      responses:
        '200':
          description: Case opened successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  opening_result:
                    type: object
                  case_price:
                    type: number
                  currency_awarded:
                    type: number
                  net_result:
                    type: number
                  new_balance:
                    type: number
        '400':
          description: Invalid case or insufficient balance
        '401':
          description: Authentication required

  # User endpoints
...